push

+++ embeddings/topwords_type.py
@@ -37,6 +37,6 @@ class Topwords:
 @dataclass
 class TopClusters:
     cluster_id: int
-    cluster_weight: float
+    cluster_weight: int
     cluster_content: List[str]
     overall_word_cosine: float
